LayoutTests/http/tests/referrer-policy-iframe/origin-when-cross-origin/cross-origin-http.https.html

<!DOCTYPE html>
<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
</head>
<body>
<script>
description("Tests the behavior of origin-when-cross-origin referrer policy when cross origin from HTTPS to HTTP.");
jsTestIsAsync = true;

if (window.internals) {
    internals.settings.setAllowDisplayOfInsecureContent(true);
    internals.setTrackingPreventionEnabled(false);
}

window.onmessage = function(event) {
    referrer = event.data.referrer;
    // Should be the origin, not the full URL, because we are cross-origin.
    shouldBeEqualToString("referrer", "https://127.0.0.1:8443/");
    if (window.internals)
        internals.setTrackingPreventionEnabled(true);
    finishJSTest();
}
</script>
<iframe src="http://localhost:8000/referrer-policy/resources/document.html" referrerpolicy="origin-when-cross-origin"></iframe>
</body>
</html>
